fix: storybook (wip) chore: moving scl-list and scl-type-list to /helpers
