+++ .storybook/preview-head.html
@@ -1,0 +1,4 @@
+<link
+  href="https://fonts.googleapis.com/css?family=Material+Icons&display=block"
+  rel="stylesheet"
+/>
